Updating the creation of properties
